@@ -159,7 +159,7 @@
     "mime": "2.4.4",
     "mini-css-extract-plugin": "0.8.0",
     "minimatch": "3.0.4",
-    "next": "9.0.5",
+    "next": "9.1.5",
     "ng-packagr": "5.7.0",
     "ngrx-store-freeze": "0.2.4",
     "npm-run-all": "^4.1.5",
