@@ -9,12 +9,12 @@ import { JsonObject } from '@angular-devkit/core';
 import {
   PHASE_DEVELOPMENT_SERVER,
   PHASE_PRODUCTION_SERVER
-} from 'next-server/constants';
+} from 'next/dist/next-server/lib/constants';
 import startServer from 'next/dist/server/lib/start-server';
 import NextServer from 'next/dist/server/next-dev-server';
 import * as path from 'path';
 import { from, Observable, of } from 'rxjs';
-import { switchMap, concatMap } from 'rxjs/operators';
+import { switchMap, concatMap, tap } from 'rxjs/operators';
 import { prepareConfig } from '../../utils/config';
 
 try {
@@ -31,7 +31,7 @@ export interface NextBuildBuilderOptions extends JsonObject {
 }
 
 export interface NextServerOptions {
-  dev: string;
+  dev: boolean;
   dir: string;
   staticMarkup: boolean;
   quiet: boolean;
@@ -77,9 +77,7 @@ function run(
           );
 
           const settings = {
-            dev: options.dev
-              ? PHASE_DEVELOPMENT_SERVER
-              : PHASE_PRODUCTION_SERVER,
+            dev: options.dev,
             dir: root,
             staticMarkup: options.staticMarkup,
             quiet: options.quiet,
@@ -93,6 +91,9 @@ function run(
             : defaultServer;
 
           return from(server(settings)).pipe(
+            tap(() => {
+              context.logger.info(`Ready on http://localhost:${settings.port}`);
+            }),
             switchMap(
               e =>
                 new Observable<BuilderOutput>(obs => {
