@@ -140,10 +140,10 @@ function createNextServerFiles(options: NormalizedSchema) {
        */
 
       /**
-       * @typedef {import('next-server/dist/server/next-server').default} Server
+       * @typedef {import('next/dist/next-server/server/next-server').default} Server
        */
 
-      const NextServer = require('next-server/dist/server/next-server').default;
+      const NextServer = require('next/dist/next-server/server/next-server').default;
       const express = require('express');
 
       const nextApp = new NextServer({
