@@ -1,6 +1,6 @@
 export const nxVersion = '*';
 
-export const nextVersion = '9.0.5';
+export const nextVersion = '9.1.5';
 export const zeitNextCss = '1.0.1';
 export const zeitNextSass = '1.0.1';
 export const zeitNextLess = '1.0.1';
