@@ -12,6 +12,7 @@ jest.mock('../../utils/app-root', () => ({ appRootPath: '/root' }));
 describe('project graph', () => {
   let packageJson: any;
   let workspaceJson: any;
+  let tsConfigJson: any;
   let nxJson: NxJson;
   let filesJson: any;
   let filesAtMasterJson: any;
@@ -70,6 +71,15 @@ describe('project graph', () => {
         util: { tags: [] }
       }
     };
+    tsConfigJson = {
+      compilerOptions: {
+        baseUrl: '.',
+        paths: {
+          '@nrwl/ui': ['libs/ui/src/index.ts'],
+          '@nrwl/util': ['libs/util/src/index.ts']
+        }
+      }
+    };
     filesJson = {
       './apps/api/src/index.ts': stripIndents`
         console.log('starting server');
@@ -88,7 +98,8 @@ describe('project graph', () => {
       `,
       './package.json': JSON.stringify(packageJson),
       './nx.json': JSON.stringify(nxJson),
-      './workspace.json': JSON.stringify(workspaceJson)
+      './workspace.json': JSON.stringify(workspaceJson),
+      './tsconfig.json': JSON.stringify(tsConfigJson)
     };
     files = Object.keys(filesJson).map(f => ({
       file: f,
