@@ -5,8 +5,15 @@ import {
   ProjectGraphNodeRecords
 } from '../project-graph-models';
 import { TypeScriptImportLocator } from './typescript-import-locator';
-import { normalizedProjectRoot } from '../../file-utils';
+import * as ts from 'typescript';
+import { appRootPath } from '../../../utils/app-root';
+import { readTsConfig } from '../../../utils/typescript';
 
+let compilerHost: {
+  host: ts.CompilerHost;
+  options: ts.CompilerOptions;
+  moduleResolutionCache: ts.ModuleResolutionCache;
+};
 export function buildExplicitTypeScriptDependencies(
   ctx: ProjectGraphContext,
   nodes: ProjectGraphNodeRecords,
@@ -20,7 +27,7 @@ export function buildExplicitTypeScriptDependencies(
       importLocator.fromFile(
         f.file,
         (importExpr: string, filePath: string, type: DependencyType) => {
-          const target = findTargetProject(importExpr, nodes);
+          const target = findTargetProject(importExpr, nodes, f.file);
           if (source && target) {
             addDependency(type, source, target);
           }
@@ -29,15 +36,39 @@ export function buildExplicitTypeScriptDependencies(
     });
   });
 
-  function findTargetProject(importExpr, nodes) {
-    return Object.keys(nodes).find(projectName => {
-      const p = nodes[projectName];
-      const normalizedRoot = normalizedProjectRoot(p);
-      return (
-        importExpr === `@${ctx.nxJson.npmScope}/${normalizedRoot}` ||
-        importExpr.startsWith(`@${ctx.nxJson.npmScope}/${normalizedRoot}#`) ||
-        importExpr.startsWith(`@${ctx.nxJson.npmScope}/${normalizedRoot}/`)
-      );
-    });
+  function findTargetProject(importExpr, nodes, filePath) {
+    compilerHost = compilerHost || getCompilerHost();
+    const { options, host, moduleResolutionCache } = compilerHost;
+
+    importExpr = importExpr.split('#')[0];
+
+    const { resolvedModule } = ts.resolveModuleName(
+      importExpr,
+      filePath,
+      options,
+      host,
+      moduleResolutionCache
+    );
+
+    // module was not found in the workspace - return
+    if (!resolvedModule) {
+      return;
+    }
+
+    return Object.keys(nodes).find(projectName =>
+      resolvedModule.resolvedFileName
+        .replace(appRootPath, '')
+        .includes(nodes[projectName].data.root)
+    );
   }
+}
+
+function getCompilerHost() {
+  const { options } = readTsConfig(`${appRootPath}/tsconfig.json`);
+  const host = ts.createCompilerHost(options, true);
+  const moduleResolutionCache = ts.createModuleResolutionCache(
+    appRootPath,
+    host.getCanonicalFileName
+  );
+  return { options, host, moduleResolutionCache };
 }
