@@ -13,6 +13,7 @@ describe('project graph', () => {
   let packageJson: any;
   let workspaceJson: any;
   let nxJson: NxJson;
+  let tsConfigJson: any;
   let filesJson: any;
   let files: FileData[];
 
@@ -48,6 +49,16 @@ describe('project graph', () => {
           sourceRoot: 'libs/shared/util/src',
           projectType: 'library'
         },
+        'shared-util-data': {
+          root: 'libs/shared/util/data',
+          sourceRoot: 'libs/shared/util/data/src',
+          projectType: 'library'
+        },
+        'lazy-lib': {
+          root: 'libs/lazy-lib',
+          sourceRoot: 'libs/lazy-lib',
+          projectType: 'library'
+        },
         api: {
           root: 'apps/api/',
           sourceRoot: 'apps/api/src',
@@ -62,7 +73,20 @@ describe('project graph', () => {
         demo: { tags: [], implicitDependencies: ['api'] },
         'demo-e2e': { tags: [] },
         ui: { tags: [] },
-        'shared-util': { tags: [] }
+        'shared-util': { tags: [] },
+        'shared-util-data': { tags: [] },
+        'lazy-lib': { tags: [] }
+      }
+    };
+    tsConfigJson = {
+      compilerOptions: {
+        baseUrl: '.',
+        paths: {
+          '@yolkai/nx-shared/util': ['libs/shared/util/src/index.ts'],
+          '@yolkai/nx-shared-util-data': ['libs/shared/util/data/src/index.ts'],
+          '@yolkai/nx-ui': ['libs/ui/src/index.ts'],
+          '@yolkai/nx-lazy-lib': ['libs/lazy-lib/src/index.ts']
+        }
       }
     };
     filesJson = {
@@ -70,7 +94,10 @@ describe('project graph', () => {
         console.log('starting server');
       `,
       './apps/demo/src/index.ts': stripIndents`
-        import * as ui from '@yolkai/ui';
+        import * as ui from '@yolkai/nx-ui';
+        import * as data from '@yolkai/nx-shared-util-data;
+
+        const s = { loadChildren: '@yolkai/nx-lazy-lib#LAZY' }
       `,
       './apps/demo-e2e/src/integration/app.spec.ts': stripIndents`
         describe('whatever', () => {});
@@ -81,9 +108,16 @@ describe('project graph', () => {
       './libs/shared/util/src/index.ts': stripIndents`
         import * as happyNrwl from 'happy-nrwl';
       `,
+      './libs/shared/util/data/src/index.ts': stripIndents`
+        export const SHARED_DATA = 'shared data';
+      `,
+      './libs/lazy-lib/src/index.ts': stripIndents`
+        export const LAZY = 'lazy lib';
+      `,
       './package.json': JSON.stringify(packageJson),
       './nx.json': JSON.stringify(nxJson),
-      './workspace.json': JSON.stringify(workspaceJson)
+      './workspace.json': JSON.stringify(workspaceJson),
+      './tsconfig.json': JSON.stringify(tsConfigJson)
     };
     files = Object.keys(filesJson).map(f => ({
       file: f,
@@ -101,17 +135,28 @@ describe('project graph', () => {
       'demo-e2e': { name: 'demo-e2e', type: 'e2e' },
       demo: { name: 'demo', type: 'app' },
       ui: { name: 'ui', type: 'lib' },
-      'shared-util': { name: 'shared-util', type: 'lib' }
+      'shared-util': { name: 'shared-util', type: 'lib' },
+      'shared-util-data': { name: 'shared-util-data', type: 'lib' },
+      'lazy-lib': { name: 'lazy-lib', type: 'lib' }
     });
-
     expect(graph.dependencies).toMatchObject({
       'demo-e2e': [
         { type: DependencyType.implicit, source: 'demo-e2e', target: 'demo' }
       ],
-      demo: expect.arrayContaining([
-        { type: DependencyType.implicit, source: 'demo', target: 'api' },
-        { type: DependencyType.static, source: 'demo', target: 'ui' }
-      ]),
+      demo: [
+        { type: DependencyType.static, source: 'demo', target: 'ui' },
+        {
+          type: DependencyType.static,
+          source: 'demo',
+          target: 'shared-util-data'
+        },
+        {
+          type: DependencyType.dynamic,
+          source: 'demo',
+          target: 'lazy-lib'
+        },
+        { type: DependencyType.implicit, source: 'demo', target: 'api' }
+      ],
       ui: [{ type: DependencyType.static, source: 'ui', target: 'shared-util' }]
     });
   });
